+++ Assets/Prefabs/Items/ExplosionParticlesSmall.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4b9a132cbc09b484f9b3efe71031bd87
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
