+++ Assets/Prefabs/Items/SmokeParticlesSmall.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2739e7addb7ef5c4398641e5413916f0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
