--- Assets/TankBoss.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c89993b8388203b4a85d7217fa3294d7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
